requirements.txt
# AI Agents System for ComfyUI - Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
